+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - core rename methods K.Util.geo.bboxContains(), bboxReverse(), bboxPlain(), bboxRound() 
 - core rename methods K.Util.geo.locBuffer(), locTransform(), locRound()
 - core simplify method K.Util.geo.bboxContains()
+- core new method L.Util.geo.polygonCentroid()
 - plugin import K.Import.geojsonToPlace()
 - core K.Util.geo.point() accept [lat,lon]